+++ src/collective/taxonomy/interfaces.py
@@ -12,7 +12,7 @@ from .i18n import CollectiveTaxonomyMessageFactory as _
 
 
 class ITaxonomySelectWidget(Interface):
-    """ Marker interface for the taxonomy select widget """
+    """Marker interface for the taxonomy select widget"""
 
 
 class IBrowserLayer(Interface):
@@ -40,37 +40,33 @@ class ITaxonomy(ITranslationDomain, IVocabularyFactory):
     def __setitem__(identifier, term):
         """ """
 
+
 # Control panel stuff
 
 
 class ITaxonomySettings(Interface):
-    """ Schema for controlpanel settings """
+    """Schema for controlpanel settings"""
 
     taxonomies = schema.List(
         title=_(u"Taxonomies"),
         value_type=schema.Choice(
             description=_(
                 u"help_taxonomies",
-                default=u"Select the taxnomies you desire to modify"
+                default=u"Select the taxonomies you desire to modify",
             ),
             required=False,
-            vocabulary='collective.taxonomy.taxonomies',
+            vocabulary="collective.taxonomy.taxonomies",
         ),
         default=[],
+        required=False,
     )
 
 
 def taxonomyDefaultValue():
-    taxonomy = api.portal.get().REQUEST.get('form.widgets.taxonomy')
+    taxonomy = api.portal.get().REQUEST.get("form.widgets.taxonomy")
     if not taxonomy:
-        return u''
+        return u""
     return taxonomy
-
-
-def get_lang_code(lang=None):
-    if lang is None:
-        lang = api.portal.get_current_language()
-    return lang.split('-', 1)[0]
 
 
 class ITaxonomyForm(Interface):
@@ -85,80 +81,73 @@ class ITaxonomyForm(Interface):
     field_title = schema.TextLine(
         title=_(u"Title"),
         description=_("Will be used for the field title as well"),
-        required=True
+        required=True,
     )
 
     field_description = schema.Text(
         title=_(u"Description"),
         description=_("Will be used for the field description as well"),
-        required=False
+        required=False,
     )
 
     default_language = schema.Choice(
         title=_(u"Default language"),
-        vocabulary='collective.taxonomy.languages',
+        vocabulary="plone.app.vocabularies.AvailableContentLanguages",
         required=True,
-        defaultFactory=get_lang_code
+        defaultFactory=api.portal.get_current_language,
     )
 
     import_file = NamedBlobFile(
-        title=_(u"Upload VDEX xml file"),
-        description=_(u" "),
-        required=False
+        title=_(u"Upload VDEX xml file"), description=_(u" "), required=False
     )
 
     import_file_purge = schema.Bool(
         title=_(u"Purge entries on upload"),
         description=_(
-            u"Check this box if you want to purge all entries "
-            u"when uploading."
+            u"Check this box if you want to purge all entries " u"when uploading."
         ),
         required=False,
     )
 
     is_required = schema.Bool(
         title=_(u"Required"),
-        description=_(
-            u"Check this box if you want the field to be required"),
-        required=True
+        description=_(u"Check this box if you want the field to be required"),
+        required=True,
     )
 
     is_single_select = schema.Bool(
         title=_(u"Single select"),
-        description=_(
-            u"Check this box if you want the field to be mono-valued"),
-        required=True
+        description=_(u"Check this box if you want the field to be mono-valued"),
+        required=True,
     )
 
     write_permission = schema.Choice(
         title=_(u"Write permission"),
-        description=_(
-            u"Write permission for the field"),
+        description=_(u"Write permission for the field"),
         required=False,
-        vocabulary='collective.taxonomy.permissions'
+        vocabulary="collective.taxonomy.permissions",
     )
 
     field_prefix = schema.ASCIILine(
         title=_(u"Field prefix"),
-        description=_(
-            u"Prefix used for behavior indexer."
-        ),
+        description=_(u"Prefix used for behavior indexer."),
         default="taxonomy_",
-        required=False
+        required=False,
     )
 
     taxonomy_fieldset = schema.TextLine(
         title=_(u"Fieldset"),
-        description=_(u"Fieldset for the taxonomy behavior field. "
-                      u"Example: 'categorization'. Use 'default' for "
-                      u"the first fieldset."),
+        description=_(
+            u"Fieldset for the taxonomy behavior field. "
+            u"Example: 'categorization'. Use 'default' for "
+            u"the first fieldset."
+        ),
         default=u"categorization",
-        required=False
+        required=False,
     )
 
 
 class ITaxonomyView(Interface):
-
     def taxonomiesForContext(self):
         """ """