+++ src/collective/taxonomy/javascripts/src/js/components/ReadOnlySubTree.js
@@ -1,48 +1,39 @@
-import React, { PropTypes } from 'react'
-import TreeView from 'react-treeview'
-import { FormattedMessage } from 'react-intl'
+import React, { PropTypes } from 'react';
+import TreeView from 'react-treeview';
+import { FormattedMessage } from 'react-intl';
 
-import ReadOnlySubTreeContainer from '../containers/ReadOnlySubTree'
+import ReadOnlySubTreeContainer from '../containers/ReadOnlySubTree';
 
-const ReadOnlySubTree = ({
-  language,
-  subnodes,
-  title
-}) => {
+const ReadOnlySubTree = ({ language, subnodes, title }) => {
   const nodeLabel = title || (
-    <FormattedMessage
-      id="untranslated"
-      defaultMessage="(Untranslated)"
-    />)
+    <FormattedMessage id="untranslated" defaultMessage="(Untranslated)" />
+  );
 
   if (subnodes.length > 0) {
     return (
-      <TreeView nodeLabel={ nodeLabel }>
-        { subnodes.map((childId) => (
+      <TreeView nodeLabel={nodeLabel}>
+        {subnodes.map(childId => (
           <ReadOnlySubTreeContainer
-            key={ childId }
-            id={ childId }
-            language={ language }
-          />))
-        }
-      </TreeView>)
+            key={childId}
+            id={childId}
+            language={language}
+          />
+        ))}
+      </TreeView>
+    );
   }
 
-  return (
-    <div className="info ro-node">
-      { nodeLabel }
-    </div>
-  )
-}
+  return <div className="info ro-node">{nodeLabel}</div>;
+};
 
 ReadOnlySubTree.propTypes = {
   language: PropTypes.string.isRequired,
   subnodes: PropTypes.array.isRequired,
-  title: PropTypes.string,
-}
+  title: PropTypes.string
+};
 
 ReadOnlySubTree.defaultProps = {
   title: ''
-}
+};
 
-export default ReadOnlySubTree
+export default ReadOnlySubTree;