+++ src/collective/taxonomy/javascripts/src/js/containers/EditableValue.js
@@ -1,16 +1,10 @@
-import { connect } from 'react-redux'
+import { connect } from 'react-redux';
 
-import { editTranslation } from '../actions'
-import EditableValue from '../components/EditableValue'
+import { editTranslation } from '../actions';
+import EditableValue from '../components/EditableValue';
 
-export const mapStateToProps = (
-  { tree: { nodes } },
-  { id, language }
-) => ({
+export const mapStateToProps = ({ tree: { nodes } }, { id, language }) => ({
   value: nodes[id].translations[language]
-})
+});
 
-export default connect(
-  mapStateToProps,
-  { editTranslation }
-)(EditableValue)
+export default connect(mapStateToProps, { editTranslation })(EditableValue);