+++ src/collective/taxonomy/javascripts/src/js/containers/LanguageSelector.js
@@ -1,17 +1,11 @@
-import { connect } from 'react-redux'
+import { connect } from 'react-redux';
 
-import { selectLanguage } from '../actions'
-import LanguageSelector from '../components/LanguageSelector'
+import { selectLanguage } from '../actions';
+import LanguageSelector from '../components/LanguageSelector';
 
-export const mapStateToProps = ({
+export const mapStateToProps = ({ defaultLanguage, languages }) => ({
   defaultLanguage,
-  languages,
-}) => ({
-  defaultLanguage,
-  languages,
-})
+  languages
+});
 
-export default connect(
-  mapStateToProps,
-  { selectLanguage },
-)(LanguageSelector)
+export default connect(mapStateToProps, { selectLanguage })(LanguageSelector);