+++ src/collective/taxonomy/javascripts/src/js/containers/MoveDownButton.js
@@ -1,10 +1,7 @@
-import { connect } from 'react-redux'
-import uid from 'uid'
+import { connect } from 'react-redux';
+import uid from 'uid';
 
-import { moveDown } from '../actions'
-import MoveDownButton from '../components/MoveDownButton'
+import { moveDown } from '../actions';
+import MoveDownButton from '../components/MoveDownButton';
 
-export default connect(
-  null,
-  { moveDown }
-)(MoveDownButton)
+export default connect(null, { moveDown })(MoveDownButton);