+++ src/collective/taxonomy/javascripts/src/js/store/configureStore.js
@@ -1,7 +1,9 @@
-import configureStoreDev from './configureStore.dev'
-import configureStoreProd from './configureStore.prod'
+import configureStoreDev from './configureStore.dev';
+import configureStoreProd from './configureStore.prod';
 
-const configureStore = process.env.NODE_ENV === 'production' ?
-  configureStoreProd : configureStoreDev
+const configureStore =
+  process.env.NODE_ENV === 'production'
+    ? configureStoreProd
+    : configureStoreDev;
 
-export default configureStore
+export default configureStore;