+++ src/collective/taxonomy/javascripts/src/js/store/middlewares.js
@@ -1,6 +1,6 @@
-import thunk from 'redux-thunk'
-import promiseMiddleware from 'redux-promise-middleware'
+import thunk from 'redux-thunk';
+import promiseMiddleware from 'redux-promise-middleware';
 
-const middlewares = [promiseMiddleware(), thunk]
+const middlewares = [promiseMiddleware(), thunk];
 
-export default middlewares
+export default middlewares;