+++ src/collective/taxonomy/javascripts/src/js/translations/fr.js
@@ -1,8 +1,9 @@
 const fr = {
   submitLabel: 'Enregistrer',
   backToSettingsLabel: 'Retour aux paramètres',
-  changesMade: 'Des changements ont été effectués. Vous devez cliquer sur le ' +
-               'bouton "Enregistrer" pour sauvegarder ces changements.',
+  changesMade:
+    'Des changements ont été effectués. Vous devez cliquer sur le ' +
+    'bouton "Enregistrer" pour sauvegarder ces changements.',
   removeNodeLabel: 'Supprimer ce nœud',
   addNodeLabel: 'Ajouter un nœud au même niveau',
   addChildNodeLabel: 'Ajouter un terme dans ce nœud',
@@ -10,7 +11,7 @@ const fr = {
   appTitle: 'Modifier les données de la taxonomie',
   emptyNodePlaceholder: 'Insérer une valeur pour ce nœud',
   untranslated: '(Non traduit)',
-  showViewTree: 'Montrer la taxonomie dans une autre langue pour comparer',
-}
+  showViewTree: 'Montrer la taxonomie dans une autre langue pour comparer'
+};
 
-export default fr
+export default fr;